Add pylint and pylintrc. Add rules to ignore. Add TODO to eval usage in SequenceGenerator
