+++ .pylintrc
@@ -1,0 +1,7 @@
+[MASTER]
+jobs=4 #number of processes to use
+[BASIC]
+good-names=nameOfYourProject #names to be considered ok
+[pre-commit-hook]
+command=custom_pylint
+disable=E0401, C0301, C0114, E0015
