+++ pyproject.toml
@@ -22,6 +22,7 @@ setuptools = "^68.2.2"
 sympy = "^1.12"
 loguru = "^0.7.2"
 dill = "^0.3.7"
+pylint = "^3.2.2"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.2"
