+++ src/sequence_generator.py
@@ -1,45 +1,8 @@
-from typing import List, Callable
+from typing import List, Callable, Dict
 from os import path
-from sympy import isprime
 import dill
 from loguru import logger
-
-
-def write_sequence(func):
-    """
-    Writes or appends a sequence to a file if a filename is provided
-    :param sequence: The sequence to write
-    """
-
-    def wrapper(self, *args, **kwargs):
-        sequence = func(self, *args, **kwargs)
-        self.sequence = sequence
-        if self.filename:
-            if path.exists(f"{self.filename}.csv") and not self.continue_sequence:
-                logger.warning(
-                    f"File {self.filename}.csv already exists. Overwriting file"
-                )
-            logger.info(f"Sequence: {sequence}")
-            with open(
-                f"{self.filename}.csv",
-                "a" if self.continue_sequence else "w",
-                encoding="UTF-8",
-            ) as f:
-                if not self.continue_sequence:
-                    # Write the type of the last element in the sequence-- this could be figured out better
-                    f.write(str(type(sequence[-1])) + "\n")
-                for s in sequence:
-                    f.write(str(s) + "\n")
-
-            if path.exists(f"{self.filename}.pickle") and not self.continue_sequence:
-                logger.warning(
-                    f"File {self.filename}.pickle already exists. Overwriting file"
-                )
-            with open(f"{self.filename}.pickle", "wb") as f:
-                f.write(dill.dumps(self))
-        return sequence
-
-    return wrapper
+from annotations import write_sequence
 
 
 class SequenceGenerator:
@@ -54,14 +17,17 @@ class SequenceGenerator:
         initial_values: List = [1, 1],
         generator: Callable = (lambda x, y: x + y),
         filename="sequence",
-        continue_sequence: bool = False,
+        **kwargs,
+        # continue_sequence: bool = False,
     ):
-        if continue_sequence:
-            with open(f"{filename}.pickle", "rb") as f:
-                self.__setstate__(dill.load(f))
-            self.continue_sequence = continue_sequence
+        options: Dict = kwargs.get("options", {})
+        self.continue_sequence = options.get("continue_sequence", False)
+        if self.continue_sequence:
+            pickle_file = f"{filename}.pickle"
+            if path.exists(pickle_file):
+                with open(f"{filename}.pickle", "rb") as f:
+                    self.__setstate__(dill.load(f))
         else:
-            self.continue_sequence = continue_sequence
             self.initial_values = initial_values
             self.generator = generator if generator else (lambda x, y: x + y)
             self.filename = filename
@@ -89,6 +55,9 @@ class SequenceGenerator:
                 # Extract the type name
                 type_line = f.readline().strip()
                 type_name = type_line.split("'")[1]
+
+                # TODO: Address this when the Generator is a class and not a lambda
+                # pylint: disable-next=eval-used
                 data_type = eval(type_name)
 
                 if tail > 0:
